little modify
