+++ pdvm_frontend/src/components/PostEdit.js
@@ -1,60 +1,93 @@
 import React, { useState } from "react"
 
-import { GetPost } from '../dataapi/Post'
+import { PostPost, PutPost } from '../dataapi/Post'
+import { PdvmButton } from '../pdvmComponents/PdvmButton'
+import PdvmInputControl  from '../pdvmComponents/PdvmInputControl'
 
-import {PostEditForm}  from '../components/PostEditForm'
-
-
-function checkCreate(pdvm, isCreate, setIsCreate) {
-  if (pdvm==='new' && !isCreate) {
-    console.log('check angekommen')
-    setIsCreate(true) 
+function useFormFields(initialValues) {
+    const [formFields, setFormFields] = useState(initialValues);
+    
+    const createChangeHandler = (key) => (
+      val: React.ChangeEvent<HTMLInputElement>,
+    ) => {
+      const value = val;
+      setFormFields((prev) => ({ ...prev, [key]: value }));
+    };
+  
+    return { formFields, createChangeHandler };
   }
-  return
-}
 
+  const initialFormFields = (formFields, post) =>{
+    formFields.id = post.id === undefined ? "" : post.id
+    formFields.userId = post.userId=== undefined ? "88" : post.userId
+    formFields.title = post.title=== undefined ? "" : post.title
+    formFields.body = post.body=== undefined ? "" : post.body
+    formFields.pdvm = post.pdvm
+    formFields.ret = post.ret
+  }
+
+/*
+Hier beginnt die default Funktion
+*/
 
 export default function PostEdit(props) {
-  const [isCreate, setIsCreate] = useState(false)
+  const { formFields, createChangeHandler } = useFormFields({
+      id: "",
+      userId: "",
+      title: "",
+      body: "",
+      pdvm: "",
+      ret: "-1",
+      })
+    
+  const [mutate_post] = PostPost(formFields)
+  const [mutate_put] = PutPost(formFields)
 
-  var post = {}
+  const [firstPdvm, setFirstPdvm] = useState(true)
+  const [isFirst, setIsFirst] = useState(true)
 
-  checkCreate(props.pdvm, isCreate, setIsCreate)
-
-  if (!isCreate) {
-    const { query, status, isLoading, data, isError, error } = GetPost(props.id)
-    console.log('Bestand GetPost',
-      query, '||',
-      status, '||',
-      isLoading, '||',
-      data, '||',
-      isError, '||',
-      error, '||'
-    )
-
-    if (isError) {
-      return <div>{error.message}</div> // error state
+  const handleSubmit = (e: React.FormEvent) => {
+    e.preventDefault();
+    if (props.isCreate) {
+      if (formFields.ret === "0" && firstPdvm) {
+        setFirstPdvm(false)
+        mutate_put(formFields)
+      } else {
+      if (firstPdvm) {
+        mutate_post(formFields)
+      } else {
+        mutate_put(formFields)
+      }}
+    } else {
+      mutate_put(formFields)
     }
-
-    if (isLoading) {
-      return <div>Daten werden geladen...</div> // loading state
-    }
-
-    post = data
-
-  } else {
-    post = props
   }
 
-  console.log('Mitte', post)
-
-  //initialFormFields(formFields, post, props)
+  if (isFirst) {
+    initialFormFields(formFields, props.post)
+    setIsFirst(false) 
+  }
 
   return (
-    <PostEditForm 
-      post={post} 
-      isCreate={isCreate}
-    />
+    <form onSubmit={handleSubmit}>
+        <PdvmInputControl
+          pdvmType='text'
+          maxLength="50"
+          type="text"
+          label="Titel"
+          value={formFields.title}
+          onChange={createChangeHandler("title")}
+        />
+        <PdvmInputControl
+          pdvmType="textaria"
+          type="text"
+          label="Artikel"
+          value={formFields.body}
+          onChange={createChangeHandler("body")}
+        />
+      <PdvmButton >
+        Speichern
+      </PdvmButton>
+    </form>
   )
-
 }
