+++ pdvm_frontend/src/pages/EditPostPage.js
@@ -1,34 +1,63 @@
-import React from 'react'
+import React, { useState } from 'react'
 
 import PostEdit from '../components/PostEdit'
+import { GetPost } from '../dataapi/Post'
 import { PdvmSection } from '../pdvmComponents/PdvmSection'
 
+function checkCreate(pdvm, isCreate, setIsCreate) {
+  if (pdvm==='new' && !isCreate) {
+    setIsCreate(true) 
+  }
+  return
+}
 
+/*
+Hier beginnt die Seite. Der Bereich zum eigentlichen Editieren befindet sich in PostEditForm.
+
+*/
 const EditPostPage = ({ match }) => {
 console.log('Match: ',match)
-  const post = {
+  let post = {
       id: match.params.id,
-      userId: match.params.userId,
-      title: match.params.title,
-      body: match.params.body,
+      userId: "-1",
+      title: "",
+      body: "",
       pdvm: match.params.pdvm,
+      ret: "-1"
     }
 
+  const [isCreate, setIsCreate] = useState(false)
+  
+  checkCreate(post.pdvm, isCreate, setIsCreate)
+  
+  if (!isCreate) {
+    const { isLoading, data, isError, error } = GetPost(post.id)
+
+    if (isError) {
+      return <div>{error.message}</div> // error state
+    }
+
+    if (isLoading) {
+      return <div>Daten werden geladen...</div> // loading state
+    }
+
+    post.id = data.id
+    post.userId = data.userId
+    post.title = data.title
+    post.body = data.body
+  }
 
   const renderPost = (
     <PostEdit 
-      id={post.id} 
-      userId={post.userId} 
-      title={post.title} 
-      body={post.body} 
-      pdvm={post.pdvm}
+      post={post} 
+      isCreate={isCreate}
     />
   )
 
-  function renderTitel(pdvm) {
-    if (pdvm === 'mod') {
+  const renderTitel = () => {
+    if (post.pdvm === 'mod') {
       return <h1>Artikel modifizieren</h1>
-    } else if (pdvm === 'new') {
+    } else if (post.pdvm === 'new') {
       return <h1>Artikel anlegen</h1>
     } else {
       return <h1>Die Art des Artikels konnte nicht ermittelt werden</h1>
@@ -37,8 +66,8 @@ console.log('Match: ',match)
 
   return (
     <PdvmSection>
-        {renderTitel(post.pdvm)}
-        {renderPost}
+      {renderTitel()}
+      {renderPost}
     </PdvmSection>
   )
 }
